Updated function to remove connection to my temp cosmos db
