+++ Functions/host.json
@@ -1,0 +1,8 @@
+{
+    "version": "2.0",
+    "extensions": {
+      "http": {
+        "routePrefix": "v1"
+      }
+    }
+  }
